Added all library test methods to Framework48 Tests

+++ Libraries/Nars-Libraries-Framework48/NarsLibraries48.cs
@@ -125,7 +125,7 @@
             /// <param name="register">Data register</param>
             /// <param name="data">Data</param>
             /// <returns>Message Data-Type</returns>
-            public result sendData(int register, long data)
+            public result sendData(int register, uint data)
             {
                 result newMessage;
